Use the correct param for exitMarket call
Problem: The call to comptroller.methods.exitMarket in the supply
dashboard uses an array of addresses as an argument. In reality,
the contract ABI accepts a single address. As a result, it is
impossible to exit a market via app.venus.io.

Solution: Replace the array with a single address.