Replace absurdly high APYs with "Infinity"
Problem: Currently most of the APYs are displayed as-is, without
any filtering. However, sometimes absurdly high values of APYs
are perfectly valid (e.g., if we have a fixed amount of XVS
distributed per day, and the amound of funds supplied/borrowed
is close to zero, we'll get numbers of the scale 10^20 or even
higher).

Solution: Add formatApy function, use it for distribution APYs and
net APYs (but not for supply/borrow APYs as wee might want to notice
the big changes there — according to our rate models such high
supply/borrow APYs cannot exist).