Adapt vault page for mobile
Problem: Due to absolute positioning, the buttons on the vault page
sometimes hide the text and inputs above.

Solution: Make a flexbox out of cards on the page. Put contents in
one block, and the button would then be another block. Use
space-between to align the card contents and the button. Allow cards
to stretch beyond 165px.