Move scripts to the top of package.json
Problem: The scripts are located between dependencies and devDependcies. This reduces their visibility and leaves the two dependency keys disorganized

Solution: Move scripts to the top of the file so that they can be referenced easily. This is convienent because developers on the project will easily be reference what scripts are available to them while working on the project.
It also has a side effect of organizing dependencies and devDependencies next to each other